requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
boto3
pyyaml
